analysis/requirements.txt
langchain
langchain-community
langchain-openai

fastapi
uvicorn
psycopg2-binary
pydantic
